% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc.quarter.R
\name{acc.quarter}
\alias{acc.quarter}
\title{Quarterly Accrual}
\usage{
acc.quarter(
  data,
  id = "dcntr_id",
  date.on = "date_on",
  open.dt,
  froz.dt,
  expmth = c(0),
  expmth.quarters = c(99),
  reg.stat = "reg",
  stars = TRUE,
  date.fmt = "\%b \%d, \%Y"
)
}
\arguments{
\item{data}{The data containing one row for each patient with on study date.}

\item{id}{Patient identifier}

\item{date.on}{Variable for on study date}

\item{open.dt}{Study open date}

\item{froz.dt}{Frozen date}

\item{expmth}{Expected patients per month for an amount of quarters set in expmth.quarters. Can be an integer 
value or vector of integers. Vector must be same length as expmth.quarters.}

\item{expmth.quarters}{Number of quarters for each expmth value. Can be integer or vector of values. 
Each vector represents the number of quarters for each expmth value. Eg. expmth = c(5,10,15) and 
expmth.quarters = c(3,3,99) would expect 5 patients accrued per month for the first 3 quarters, 
10 patients accrued per month for quarters 4-6, and 15 patients per month thereafter.}

\item{reg.stat}{Registration status for accrual table. "reg" for registered, "pre" for pre-registered, 
and "rand" for randomized.}

\item{stars}{Option to add asterisks to the first 2 quarter start times and the last quarter stop time for 
comments or footnotes. Set to FALSE if no asterisks are desired.}

\item{date.fmt}{Format for dates in the table. Default "\%b \%d, \%Y".}
}
\value{
This function creates a quarterly accrual table.
}
\description{
This function creates the quarterly accrual table for Alliance DSMB reports.
}
\examples{

crtlibn(d = "A041501")

x <- acc.quarter(casecrse, open.dt = "6/15/2018", froz.dt = "3/20/2020", expmth = 8)
x

y <- acc.quarter(casecrse, open.dt = "6/15/2018", froz.dt = "3/20/2020", expmth = c(5,10,15),
 expmth.quarters = c(2,2,99))
y

}
\author{
Sawyer Jacobson
}
