% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accgph.R
\name{accgph}
\alias{accgph}
\title{Create monthly accrual graph}
\usage{
accgph(
  data,
  protnum,
  id = "dcntr_id",
  date.on,
  open.dt,
  froz.dt,
  expmth = 0,
  y.breaks = round(y.lim/10, -1),
  exp.pts,
  y.lim = nrow(one)
)
}
\arguments{
\item{data}{The dataset with the patients you want to use. Must contain patient IDs and patient "on date"}

\item{protnum}{The study number as a character.}

\item{id}{The patient ID variable in quotes used in the dataset}

\item{date.on}{The variable that contains the patient "on date". Entered as character.}

\item{open.dt}{The study open date. Alternatively, the lower limit for date you want to use for the start of the graph. Entered in "mm/dd/yyyy" format.}

\item{froz.dt}{The upper limit for the patient "on dates". Entered in "mm/dd/yyyy" format.}

\item{expmth}{The number of expected patients accrued per month. Available in protocol.}

\item{y.breaks}{The distance between each y-axis break. Default is y.lim/10 rounded to the nearest 10.}

\item{exp.pts}{The number of patient expected to be accrued for the study}

\item{y.lim}{Limit for y-axis. Default is the number of rows in the given dataset.}
}
\value{
This function creates a month by month accrual graph.
}
\description{
This is a function to create accrual graphs for trials. Similar output to the \%accgph() macro.
}
\examples{

#With data loaded for A041501

crtlibn(d = 'A041501')

accrual <- accgph(protnum = "A041501", 
    data = crse, id = "dcntr_id", 
    date.on = "date_on",
    open.dt = "06/15/2018",
    froz.dt = format(Sys.Date(), "\%m/\%d/\%Y"), 
    expmth = 5, exp.pts = 324) 
accrual

}
\author{
Sawyer Jacobson

Adam Pettinger
}
