% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aerpt.summary.R
\name{aerpt.summary}
\alias{aerpt.summary}
\title{Create summary of AEs for Alliance DSMB reports}
\usage{
aerpt.summary(
  cytox,
  id,
  grade.cutoff = 1,
  rel.cutoff = 0,
  rel.include = 1,
  byvar = "arm",
  arm.include = 1,
  arm.labels = NULL,
  toxvar = "toxicity",
  relvar = "rel_smed",
  gradevar = "grade",
  bodysysvar = "bodysys"
)
}
\arguments{
\item{cytox}{The toxicity dataset containing multiple observations per patient with one observation per AE > grade 0}

\item{id}{The patient ID variable used in the dataset}

\item{grade.cutoff}{Grade cutoff for minimum AE grade. Inclusive}

\item{rel.cutoff}{Attribution cutoff for minimum AE attribution. Inclusive}

\item{rel.include}{Value of 0 or 1. If 0, "Regardless of Attribution" printed in header. If 1, "At least possibly related" printed in header}

\item{byvar}{Variable used to separate summary by arm. If not present, "All Patients" printed for arm. Set to NULL if calculations are desired for all patients together.}

\item{arm.include}{Include "Arm" in table (1) or not (0). Default is 1.}

\item{arm.labels}{Option to include additional Arm labels in the table header. Add as vector of characters.}

\item{toxvar}{Variable containing numeric toxicity codes.}

\item{relvar}{Variable containing numeric relation values 1-5.}

\item{gradevar}{Grade variable, numeric.}

\item{bodysysvar}{Body system variable, character.}
}
\value{
This function creates a summary of the AE events for the Alliance DSMB report
}
\description{
This function is design to replicate aerpt_alliance_macro.sas AE summary output for Alliance DSMB reports.
}
\examples{

#With data loaded for A041501

crtlibn(d = 'A041501')

all.summary <- aerpt.summary(cytox, id="dcntr_id", byvar="arm", 
grade.cutoff = 3, rel.cutoff = 0, rel.include = 0)
all.summary$table

}
\author{
Sawyer Jacobson

Adam Pettinger
}
