% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bodysys_formats.R
\docType{data}
\name{bodysys_formats}
\alias{bodysys_formats}
\title{Body System Formats Data}
\format{
A data frame with mappings from numeric body system codes to descriptions.
}
\usage{
bodysys_formats
}
\description{
Body System Formats Data
}
\keyword{datasets}
