% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eot.table.R
\name{eot.table}
\alias{eot.table}
\title{End of Treatment Table}
\usage{
eot.table(data, id = "dcntr_id", endatrsn = "endatrsn", arm = "arm")
}
\arguments{
\item{data}{A dataset with 1 row for each patient accrued.}

\item{id}{Patient ID variable}

\item{endatrsn}{End of treatment reason variable. May be numeric or factored.}

\item{arm}{Arm variable.}
}
\value{
This function returns an end of treatment table formatted for the Alliance DSMB report.
}
\description{
This function creates the end of treatment table for Alliance DSMB reports
}
\details{
If endatrsn column is a factor, the given values will be used. However, if the column is 
numeric, the values will be mapped to pre-specified descriptions.

Value map for numeric input:

NA = On Treatment

1 = Treatment Completed Per Protocol

2 = Patient Withdrawal/Refusal After Beginning Protocol Therapy

3 = Adverse Events/Side Effects/Complications

4 = Disease Progression, Relapse During Active Treatment

5 = Alternative Therapy

6 = Patient Off-Treatment For Other Complicating Disease

7 = Death On Study

8 = Other

10 = Disease Progression Before Active Treatment

24 = Patient Withdrawal/Refusal Prior To Beginning Protocol Therapy
}
\examples{

crtlibn(d = "A041501")

x = eot.table(casecrse, arm = "curr_arm")
x$table

}
\author{
Sawyer Jacobson
}
