% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot.R
\name{forest_plot}
\alias{forest_plot}
\title{Forest Plot}
\usage{
forest_plot(
  data,
  stat,
  time,
  byvar,
  covars,
  ref = as.character(levels(data2$arm)[1]),
  labels = NULL,
  graph.pos = 3,
  xaxis.breaks = 0.25,
  title = "Hazard Rates",
  x.lab = "Hazard Ratios",
  plot.include = TRUE,
  x.max = max(values$upper, na.rm = TRUE),
  ...
)
}
\arguments{
\item{data}{A dataset containing the necessary survival variables (event, time, other covariables).}

\item{stat}{The event survival variable, numeric, with 1 coded as an event and 0 a censor.}

\item{time}{The time survival variable, numeric.}

\item{byvar}{The by variable with 2 distinct levels.}

\item{covars}{A vector of categorical covariates that are desired for analysis.}

\item{ref}{Reference by variable level (optional). Use this if specific by variable level is desired as the reference.}

\item{labels}{An optional vector, soon to be list, of labels for the covariates in the 'covars' variable. Currently must be the same length as the 'covars' vector.}

\item{graph.pos}{Column position for the graph.}

\item{xaxis.breaks}{Distance between the x-axis breaks of the forest plot.}

\item{title}{Title for the plot.}

\item{x.lab}{Label for x-axis of plot.}

\item{plot.include}{Logical to have plot return from function call.}

\item{x.max}{Max value for x-axis. CI bands for intervals that pass this value will have an arrow on the right indicating a continuing band.}

\item{...}{Additional arguments for the 'forestplot' function.}
}
\value{
This function returns a list of the forestplot produced and tables necessary to reproduce that plot.
}
\description{
This function creates the data for and plots a forest plot comparing hazard rates for categorical variables by a 2 level by variable.
}
\examples{

library(arsenal)
data(mockstudy)
mockstudy \%>\%
 mutate(fu.stat = fu.stat - 1) \%>\%
 filter(arm != "A: IFL" & age.ord != "10-19" & race != "Other") \%>\%
 forest_plot(data = ., stat = "fu.stat", time = "fu.time", byvar = "arm", 
             covars = c("sex", "age.ord"), title = "Overall Survival", 
             labels = c("Gender", "Age Groups"), 
             xaxis.breaks = .5)

}
