% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irbtox_listing.R
\name{irbtox_listing}
\alias{irbtox_listing}
\title{Create listing of AEs for Mayo DSMB reports}
\usage{
irbtox_listing(
  cytox,
  id = "dcntr_id",
  byvar = "arm",
  grade.term = "grade",
  relation.term = "rel_smed",
  tox.term = "toxicity"
)
}
\arguments{
\item{cytox}{The toxicity dataset containing multiple observations per patient with one observation per AE > grade 0}

\item{id}{The patient ID variable used in the dataset}

\item{byvar}{Variable used to separate summary by arm. If not present, "All Patients" printed for arm}

\item{grade.term}{Character. The variable that contains the AE grade}

\item{relation.term}{Character. The variable that containst the AE attribution}

\item{tox.term}{Character. The variable that contains the toxicity (MEDRA) code}
}
\value{
This function creates a listing of the AE events for the Mayo DSMB report
}
\description{
This function is design to replicate irbtoxrep.sas macro AE listing output for Mayo DSMB reports.
}
\examples{

#With data loaded for A041501

crtlibn(d = 'A041501')

irb.listing <- irbtox_listing(cytox = cytox, id = "dcntr_id", byvar = "arm", tox.term = "toxicity")
irb.listing$table

}
\author{
Sawyer Jacobson

Adam Pettinger
}
