% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irbtox_summary.R
\name{irbtox_summary}
\alias{irbtox_summary}
\title{Create summary of AEs for Mayo DSMB reports}
\usage{
irbtox_summary(
  cytox,
  id = "dcntr_id",
  byvar = "arm",
  grade.term = "grade",
  relation.term = "rel_smed",
  tox.term = "toxicity",
  prot_num = "",
  s_desc = ""
)
}
\arguments{
\item{cytox}{The toxicity dataset containing multiple observations per patient with one observation per AE > grade 0}

\item{id}{The patient ID variable used in the dataset}

\item{byvar}{Variable used to separate summary by arm. If not present, "All Patients" printed for arm}

\item{grade.term}{Character. The variable that contains the AE grade}

\item{relation.term}{Character. The variable that containst the AE attribution}

\item{tox.term}{Character. The variable that contains the toxicity (MEDRA) code}

\item{prot_num}{The study protocol number. Found in protref$dc_num}

\item{s_desc}{Short description of the study. Found in protref$s_desc.}
}
\value{
This function returns a list of 1: Protocol Number and evaluable patients per Arm as a string, 2: Summary AE table, and 3: Total number of evaluable patients.
}
\description{
This function is design to replicate irbtoxrep.sas macro AE summary output for Mayo DSMB reports.
}
\examples{

#With data loaded for A041501

crtlibn(d = 'A041501')

irb.summary <- irbtox_summary(cytox = cytox, id = "dcntr_id", byvar = "arm", tox.term = "toxicity",
 prot_num = protref$dc_num, s_desc = protref$shrt_ttl_m)
irb.summary$table

}
\author{
Sawyer Jacobson

Adam Pettinger
}
