% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km.plot.R
\name{km.plot}
\alias{km.plot}
\title{KM plotting function}
\usage{
km.plot(
  data,
  stat,
  time,
  arm = "arm",
  x.lab = "Time (Months)",
  x.time = NULL,
  y.lab = "\% Event-Free",
  ref = as.character(levels(data2$arm)[1]),
  x.max = max(data2$time),
  x.by = 12,
  surv.legend.labs = levels(data2$arm2),
  table.legend.labs = levels(data2$arm2),
  color = RColorBrewer::brewer.pal(8, "Set1"),
  lty = 1,
  table.font.size = 9,
  y.scale = "percent",
  one.sided = FALSE,
  risk.table = TRUE,
  table.include = TRUE,
  pval.include = TRUE,
  pval = NULL,
  pval.loc = NULL,
  size = 0.5,
  risk.table.height = 0.175,
  surv.plot.height = 0.825,
  table.x.min = x.max/5,
  table.x.max = x.max/1.5,
  x.scale = 1,
  surv.order = NULL,
  surv.legend.title = "Arm: ",
  legend.font = 9,
  theme = survminer::theme_survminer(),
  table.loc = "topright",
  ...
)
}
\arguments{
\item{data}{The data.}

\item{stat}{The survival status variable used in survfit()}

\item{time}{The survival time variable used in survfit()}

\item{arm}{By variable used in survfit()}

\item{x.lab}{Label for x-axis of KM plot}

\item{x.time}{Specific unit for the survival time variable in x-axis label, character (eg. "Days", "Months", "Years").}

\item{y.lab}{Label for y-axis of KM plot}

\item{ref}{Specify reference arm for HR values in survival table.}

\item{x.max}{Max value for the x-axis}

\item{x.by}{By value for tick marks on the x-axis based on the scaled time variable.}

\item{surv.legend.labs}{KM plot legend arm labels}

\item{table.legend.labs}{Survival table arm labels}

\item{color}{Color for KM plot lines}

\item{lty}{Type of lines for KM plot}

\item{table.font.size}{Font size for the survival table}

\item{y.scale}{Scale for y-axis. Options are "default" (decimals, 0.20) and "percent" (percent, 20\%)}

\item{one.sided}{TRUE for one-sided p-value, FALSE for two-sided p-value}

\item{risk.table}{Include risk table (TRUE/FALSE)}

\item{table.include}{Include survival table (TRUE/FALSE)}

\item{pval.include}{Include p-value. if TRUE and pval = NULL, logrank p-value from survfit() is included}

\item{pval}{User specified p-value. This can be used to specify p-value used if something different than logrank is desired. Can be character or numeric.}

\item{pval.loc}{Vector of length 2 to specify the position for the p-value if necessary.}

\item{size}{Overall size of the plot?}

\item{risk.table.height}{Percent of plot size taken up by the risk table.}

\item{surv.plot.height}{Percent of plot size taken up by the KM plot.}

\item{table.x.min}{Minimum x-axis value for the survival table if table.loc = "topright".}

\item{table.x.max}{Maximum x-axis value for the survival table if table.loc = "bottomleft".}

\item{x.scale}{Numeric variable to allow for scaling of the time variable. Will divide the time variable by the value given.}

\item{surv.order}{Character vector. Allows for the display order of the variables in the legend, plot, survival table, and risk table to be specified. Must include all levels of the "arm" variable.}

\item{surv.legend.title}{Legend title to appear at the top of the plot (i.e. "Arm: ", "Strata: ').}

\item{legend.font}{Font size for the legend.}

\item{theme}{Desired theme for the KM plot and risk table.}

\item{table.loc}{Character. Specify where the survival table should go inside the plot. Currently support "topright" and "bottomleft".}

\item{...}{Additional arguments passed onto ggsurvplot()}
}
\value{
This function returns a publication quality KM curve.
}
\description{
This is a function to consistently plot aesthetically pleasing KM plots using survminer::ggsurvplot() that include risk table below and survival table within.
}
\examples{

library(arsenal)
data(mockstudy)
mockstudy \%>\% 
    km.plot(data = ., stat = "fu.stat", time = "fu.time", arm = "arm", 
            table.loc = "topright", x.scale = 30.44)

}
