% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_table.R
\name{surv_table}
\alias{surv_table}
\title{Create survfit table}
\usage{
surv_table(data, time, event, by)
}
\arguments{
\item{data}{data frame that contains time to event, event, and by variables required to create survfit object}

\item{time}{Time variable}

\item{event}{Event variable}

\item{by}{By variable}
}
\value{
This function returns a data.frame containing a column for each level of the by variable and the pvalue with rows "N", "Events" (# events),
 "Median" (median survival time), "95\% CI" (for median survival time).
}
\description{
Use this to create a table for survival analysis. Creates a data.frame containing sample size, number of events, median survival time, and 95\% CI between by groups.
}
