% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_tableby.R
\name{tidy_tableby}
\alias{tidy_tableby}
\title{Tidy Word Tableby}
\usage{
tidy_tableby(x)
}
\arguments{
\item{x}{Dataframe of a summary() of tableby object}
}
\value{
This function outputs a tidy version of a dataframe summary(tableby) object to be output in word.
}
\description{
Tidy Word Tableby
}
