% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tox_codes.R
\docType{data}
\name{tox_codes}
\alias{tox_codes}
\title{Toxicity Codes Data}
\format{
A data frame with mappings from numeric toxicity codes to CTCAE descriptions
}
\usage{
tox_codes
}
\description{
Toxicity Codes Data
}
\keyword{datasets}
